v0.4.0

+++ Classes/Core/FanRSA.m
@@ -296,7 +296,7 @@ static NSData *base64_decode(NSString *str){
                                &outlen
                                );
         if (status != 0) {
-            NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %ld", (long)status);
             ret = nil;
             break;
         }else{
@@ -377,7 +377,7 @@ static NSData *base64_decode(NSString *str){
                                &outlen
                                );
         if (status != 0) {
-            NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %ld", (long)status);
             ret = nil;
             break;
         }else{
